99/12/07:
- Fixed some bugs
- Reduced extra endpoints

+++ src/main/java/ir/darkdeveloper/sma/Post/Models/CommentModel.java
@@ -27,7 +27,7 @@ public class CommentModel  {
     @Column(nullable = false)
     private String content;
 
-    private int likes;
+    private Long likes;
 
     @ManyToOne
     private PostModel post;
